+++ packages/ns-build/server.js
@@ -51,10 +51,6 @@ const parseEvent = (event) => {
     const parsedEvent = Object.assign(event);
     parsedEvent.path = parsedEvent.rawPath;
     parsedEvent.headers.host = parsedEvent.headers['x-forwarded-host'];
-    parsedEvent.headers.referer =
-        parsedEvent.headers['x-forwarded-proto'] +
-            '://' +
-            parsedEvent.headers['x-forwarded-host'];
     return parsedEvent;
 };
 // Convert route file path to regex
@@ -276,9 +272,9 @@ const handler = (event, context, callback) => {
     /* If an image is requested, redirect to the corresponding S3 bucket. */
     if (imageTypes.some(type => parsedEvent.path.includes('.' + type))) {
         const response = {
-            statusCode: 301,
+            statusCode: 302,
             headers: {
-                Location: parsedEvent.headers.referer + '/assets' + parsedEvent.path,
+                Location: 'https://' + parsedEvent.headers.host + '/assets' + parsedEvent.path,
             },
         };
         return callback(null, response);